+++ src/step2-index.js
@@ -3,7 +3,7 @@
  * 노드 환경에서 사용하는 readline 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
  */
 import './web/css/index.css';
-import WebLottoController from './web/WebLottoController';
+import WebLottoController from './web/controller/WebLottoController';
 
 const webLottoController = new WebLottoController();
